+++ src/SurrealDB.QueryBuilder/DataModels/Geometry/Point.cs
@@ -6,47 +6,103 @@ namespace SurrealDB.QueryBuilder.DataModels.Geometry;
 /// </summary>
 public readonly struct Point : IGeometry, IEquatable<Point>
 {
-    public readonly decimal X { get; }
+    /// <summary>
+    /// Gets the X coordinate of this <see cref="Point"/> structure, representing the longitude.
+    /// </summary>
+    /// <returns>The X coordinate of this <see cref="Point"/> structure, representing the longitude.</returns>
+    public readonly double X { get; }
 
-    public readonly decimal Y { get; }
+    /// <summary>
+    /// Gers the Y coordinate of this <see cref="Point"/> structure, representing the latitude.
+    /// </summary>
+    /// <returns>The Y coordinate of this <see cref="Point"/> structure, representing the latitude.</returns>
+    public readonly double Y { get; }
 
-    public Point(decimal x, decimal y)
+    /// <summary>
+    /// Initializes a new instance of the <see cref="Point"/> structure with the specified coordinates.
+    /// </summary>
+    /// <param name="x">The X coordinate (longitude).</param>
+    /// <param name="y">The Y coordinate (latitude).</param>
+    public Point(double x, double y)
     {
         X = x;
         Y = y;
     }
 
-    public void Deconstruct(out decimal x, out decimal y)
+    /// <summary>
+    /// Deconstructs this <see cref="Point"/> into its <see cref="X"/> and <see cref="Y"/> components.
+    /// </summary>
+    /// <param name="x">The X coordinate (longitude).</param>
+    /// <param name="y">The Y coordinate (latitude).</param>
+    public void Deconstruct(out double x, out double y)
     {
         x = X;
         y = Y;
     }
 
-    public static implicit operator Point((decimal longitude, decimal latitude) coordinates)
+    /// <summary>
+    /// Implicitly converts a tuple containg a longitude and latitude value to a <see cref="Point"/> structure.
+    /// </summary>
+    /// <param name="coordinates">The tuple containing the longitude and latitude values respectively.</param>
+    public static implicit operator Point((double longitude, double latitude) coordinates)
         => new(coordinates.longitude, coordinates.latitude);
 
-    public static implicit operator (decimal x, decimal y)(Point point)
+    /// <summary>
+    /// Implicitly converts a <see cref="Point"/> structure to a tuple containing a longitude and latitude value.
+    /// </summary>
+    /// <param name="point">The <see cref="Point"/> structure to convert from.</param>
+    public static implicit operator (double x, double y)(Point point)
         => (point.X, point.Y);
 
     /// <inheritdoc/>
     public SchemalessObject ToGeoJson()
         => new() { { "type", nameof(Point) }, { "coordinates", CoordinatesOnly() } };
 
-    internal decimal[] CoordinatesOnly()
+    /// <summary>
+    /// Returns an array containing the <see cref="X"/> and <see cref="Y"/> coordinates of this <see cref="Point"/>.
+    /// </summary>
+    /// <returns>An array containing the <see cref="X"/> and <see cref="Y"/> coordinates of this <see cref="Point"/>.</returns>
+    internal double[] CoordinatesOnly()
         => new[] { X, Y };
 
+    /// <summary>
+    /// Returns a value that indicates whether this instance is equal to a specified <see cref="Point"/>.
+    /// </summary>
+    /// <param name="value">The <see cref="Point"/> to compare to this instance.</param>
+    /// <returns><see langword="true"/> if <paramref name="value"/> has the same <see cref="X"/> and <see cref="Y"/> as this instance; otherwise, <see langword="false"/>.</returns>
     public bool Equals(Point value)
         => X == value.X && Y == value.Y;
 
+    /// <summary>
+    /// Returns a value that indicates whether this instance is equal to a specified object.
+    /// </summary>
+    /// <param name="value">The object to compare to this instance.</param>
+    /// <returns><see langword="true"/> if <paramref name="value"/> is a <see cref="Point"/> that has the same <see cref="X"/> and <see cref="Y"/> as this instance; otherwise, <see langword="false"/>.</returns>
     public override bool Equals(object? value)
         => value is Point point && Equals(point);
 
+    /// <summary>
+    /// Returns the hash code for this instance.
+    /// </summary>
+    /// <returns>A 32-bit signed integer hash code.</returns>
     public override int GetHashCode()
         => HashCode.Combine(X, Y);
 
+    /// <summary>
+    /// Returns a value indicating wether this instance is equal to a specified <see cref="Point"/>.
+    /// </summary>
+    /// <param name="left"><see cref="Point"/> to compare.</param>
+    /// <param name="right"><see cref="Point"/> to compare.</param>
+    /// <returns><see langword="true"/> if <paramref name="left"/> and <paramref name="right"/> have the same <see cref="X"/> and <see cref="Y"/>; otherwise, <see langword="false"/>.</returns>
     public static bool operator ==(Point left, Point right)
         => Equals(left, right);
 
+    /// <summary>
+    /// Returns a value indicating wether this instance is not equal to a specified <see cref="Point"/>.
+    /// </summary>
+    /// <param name="left"><see cref="Point"/> to compare.</param>
+    /// <param name="right"><see cref="Point"/> to compare.</param>
+    /// <return><see langword="true"/> if <paramref name="left"/> and <paramref name="right"/> do not have the same <see cref="X"/> and <see cref="Y"/>; otherwise, <see langword="false"/>.</return>
     public static bool operator !=(Point left, Point right)
         => !Equals(left, right);
 }
